backend/requirements.txt
# Python packages
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
redis==5.0.0
python-dotenv==1.0.0
pydantic==2.3.0
aiohttp==3.8.5
openai==0.28.0
python-jose==3.3.0
passlib==1.7.4